% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabSolution-pkg.R
\name{eqval}
\alias{eqval}
\title{Quantity of a substance that interacts with one mole of hydrogen ions (in an acid-base reaction) or electrons (in a redox reaction)}
\usage{
eqval(fw, n)
}
\arguments{
\item{fw}{Formula weight (Molecular weight or molar mass) of the solute}

\item{n}{The number of equivalents that changes by ionization}
}
\value{
fw / n
}
\description{
Equivalent: In a chemical reaction, an equivalent is the amount of a substance that reacts with (or is equivalent to) an arbitrary amount (typically one mole) of another substance
}
\details{
What is the gram equivalent weight of phosphoric acid, H3PO4?(fw= molecular weight in g/mol, n= number of equivalents)
}
\examples{
fw=127
n= 3
eqval(127,3)
}
\author{
Pejman Morovat <pejmanmorovat@yahoo.com>
}
