% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabSolution-pkg.R
\name{mgrams}
\alias{mgrams}
\title{Determination of the amount of solute in grams for preparing the solution  when the unit of solution concentration is the molarity}
\usage{
mgrams(V, molarity, fw)
}
\arguments{
\item{V}{Volume of solution (ml)}

\item{molarity}{moles of solute / liters of solution (mol/L).}

\item{fw}{Formula weight or molar mass of the solute}
}
\value{
v * molarity * fw / 1000
}
\description{
Determination of the amount of solute in grams for preparing the solution  when the unit of solution concentration is the molarity
}
\details{
To make 250 ml of an aqueous solution with [NaCl] = 0.8 M. What mass of the solute would we need to make this solution? the molecular weight of sodium chloride= 58.44 g/mol
}
\examples{
v=250
normality=0.8
fw=58.44
mgrams(250,0.8,58.44)
}
\author{
Pejman Morovat <pejmanmorovat@yahoo.com>
}
