% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabSolution-pkg.R
\name{molarity}
\alias{molarity}
\title{Molarity: Number of moles of solute per one liter of solutions (Solvent+Solute)}
\usage{
molarity(p, d, fw)
}
\arguments{
\item{p}{Purity of solution. Percentage purity of a substance can be calculated by dividing the mass of the pure chemical by the total mass of the sample, and then multiplying this number by 100}

\item{d}{Density of solution. To determine the solution's density, weigh a precisely measured volume of your solution, and divide the mass of solution by the volume of solution. You must measure the volume of solution to obtain the density}

\item{fw}{Formula weight (Molecular weight or molar mass) of the solute. If the mass of the unknown solute is known, then its molar mass can be calculated using the formula: Molar mass of solute = mass of solute/number of moles of solute}
}
\value{
10 * p (the percentage purity) * d (density of solution(g/cm3)) / fw (Molar mass(g/mol))
}
\description{
Molarity: An amount of solute in one unit of a solution is known as the molar concentration. The unit of measurement is mol/L, mol/dm3, or mol/m3. In molar concentration, also known as molarity, the unit M refers to the molar.
}
\details{
What is the molarity of concentrated sulfuric acid if its purity is 96 percent and its density is 1.84 grams per liter? The molecular weight of HSO4 is 97.064 g/mol.
}
\examples{
p=96
d=1.84
fw=97.064
molarity(96,1.84,97.064)
}
\author{
Pejman Morovat <pejmanmorovat@yahoo.com>
}
