% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabSolution-pkg.R
\name{ngrams}
\alias{ngrams}
\title{Determination of the amount of solute in grams for preparing the solution  when the unit of solution concentration is normality}
\usage{
ngrams(v, normality, eqval)
}
\arguments{
\item{v}{Volume of solution (ml)}

\item{normality}{Normality of solution. Number of gram equivalents/volume of solution in liters (eq/L)}

\item{eqval}{Equivalent: In a chemical reaction, an equivalent is the amount of a substance that reacts with (or is equivalent to) an arbitrary amount (typically one mole) of another substance}
}
\value{
v * normality * eqval / 1000
}
\description{
Determination of the amount of solute in grams for preparing the solution  when the unit of solution concentration is normality
}
\details{
To make 150 ml of an aqueous solution with [NaCl] = 10 N. What mass of the solute would we need to make this solution?
}
\examples{
v=150
normality=0.8
eqval=1
ngrams(150,10,1)
}
\author{
Pejman Morovat <pejmanmorovat@yahoo.com>
}
