% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabSolution-pkg.R
\name{normality}
\alias{normality}
\title{Normality: Defied as the number of gram of equivalent of solute per volume of solution (eq/L)}
\usage{
normality(p, d, eqval)
}
\arguments{
\item{p}{Purity of solution. Percentage purity of a substance can be calculated by dividing the mass of the pure chemical by the total mass of the sample, and then multiplying this number by 100}

\item{d}{Density of solution. To determine the solution's density, weigh a precisely measured volume of your solution, and divide the mass of solution by the volume of solution. You must measure the volume of solution to obtain the density}

\item{eqval}{Equivalent: In a chemical reaction, an equivalent is the amount of a substance that reacts with (or is equivalent to) an arbitrary amount (typically one mole) of another substance}
}
\value{
10 * p * d / eqval
}
\description{
Normality is the number of equivalents per liter of solution.
}
\details{
What is the normality of a 96 per cent solution of H2SO4 of specific gravity 1.84 g/mL?
}
\examples{
p=96
d=1.84
eqval=49.039
normality(96,1.84,49.039)
}
\author{
Pejman Morovat <pejmanmorovat@yahoo.com>
}
