% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabSolution-pkg.R
\name{normality2}
\alias{normality2}
\title{It's another function for the determination of the normality of the solution.}
\usage{
normality2(molarity, n)
}
\arguments{
\item{molarity}{Molarity of solution (mol/L). moles of solute/liters of solution}

\item{n}{The number of equivalents that changes by ionization}
}
\value{
molarity * n
}
\description{
Normality= Molarity * Equivalent mass (g/eq)
}
\details{
What is the normality of Sulfuric Acid?(molarity of H2SO4 96 percent= 18.01)
}
\examples{
molarity=18.01
n=2
normality2(18.01,2)
}
\author{
Pejman Morovat <pejmanmorovat@yahoo.com>
}
