% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabSolution-pkg.R
\name{vol}
\alias{vol}
\title{Using the following equation, you can figure out the concentration or volume of the concentrated or diluted solution}
\usage{
vol(V, molarity, m)
}
\arguments{
\item{V}{is the volume of the dilute solution}

\item{molarity}{is the concentration in molarity of the dilute solution (after more solvent has been added)}

\item{m}{is the concentration in molarity (mol/L) of the concentrated solution}
}
\value{
V * molarity / m
}
\description{
m * vol = molarity * v --> , vol is the volume of the concentrated solution
}
\details{
Prepare 100 mL of 1.0 M hydrochloric acid from concentrated (12.1 M) hydrochloric acid.
}
\examples{
v=100
molarity=1
m=12
vol(100,1,12)
}
\author{
Pejman Morovat <pejmanmorovat@yahoo.com>
}
